create separate class for configuration and update references
